admin and superadmin check added

+++ migrations/20231030205921-create-order.js
@@ -10,7 +10,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       total_price: {
-        type: Sequelize.FLOAT(11, 10)
+        type: Sequelize.FLOAT
       },
       AddressId: {
         type: Sequelize.INTEGER
